sync

+++ go/vt/schemamanager/local_controller_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2017 Google Inc.
+Copyright 2019 The Vitess Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -25,9 +25,9 @@ import (
 	"strings"
 	"testing"
 
-	"golang.org/x/net/context"
+	"context"
 
-	querypb "github.com/youtube/vitess/go/vt/proto/query"
+	querypb "vitess.io/vitess/go/vt/proto/query"
 )
 
 func TestLocalControllerNoSchemaChanges(t *testing.T) {
@@ -55,15 +55,15 @@ func TestLocalControllerOpen(t *testing.T) {
 	controller := NewLocalController("")
 	ctx := context.Background()
 
-	if err := controller.Open(ctx); err == nil {
-		t.Fatalf("Open should fail, no such dir")
+	if err := controller.Open(ctx); err == nil || !strings.Contains(err.Error(), "no such file or directory") {
+		t.Fatalf("Open should fail, no such dir, but got: %v", err)
 	}
 
-	schemaChangeDir, err := ioutil.TempDir("", "localcontroller-test")
+	schemaChangeDir, _ := ioutil.TempDir("", "localcontroller-test")
 	defer os.RemoveAll(schemaChangeDir)
 
 	// create a file under schema change dir
-	_, err = os.Create(path.Join(schemaChangeDir, "create_test_table.sql"))
+	_, err := os.Create(path.Join(schemaChangeDir, "create_test_table.sql"))
 	if err != nil {
 		t.Fatalf("failed to create sql file, error: %v", err)
 	}
